Hydrate views even if registered after DOM nodes appear
