+++ src/blocks/interactive-child/block.json
@@ -21,5 +21,5 @@
 	"textdomain": "block-hydration-experiments",
 	"editorScript": "file:./index.js",
 	"style": "file:./style-index.css",
-	"viewScript": "file:./frontend.js"
+	"viewScript": "file:./register-view.js"
 }
