+++ src/blocks/interactive-parent/block.json
@@ -43,5 +43,5 @@
 	"textdomain": "block-hydration-experiments",
 	"editorScript": "file:./index.js",
 	"style": "file:./style-index.css",
-	"viewScript": "file:./frontend.js"
+	"viewScript": "file:./register-view.js"
 }
